replace shelljs.exec with child_process.exec/execSync/spawnSync
closes #14

This replaces all instances of shelljs.exec with the appropriate method
from child_process (that is, either exec, execSync or spawnSync). This
implicates that we drop support for Node.js v0.10 and below, because
spawnSync and execSync have been introduced in v0.12.

shell.js has been kept as a dependency for shelljs.which.

Related minor changes:
- specify node version in package.json
- configure node v0.12 for travis
- document Node.js version support in README.md